requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML and Data Science
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2

# Vector Search and Embeddings
faiss-cpu==1.7.4
# Use faiss-gpu==1.7.4 if you have GPU support

# OpenAI Integration
openai==1.3.8

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For enhanced functionality
python-multipart==0.0.6
python-dotenv==1.0.0
